Move out tags to gtags.hpp & add protection for own::Mat
