+++ modules/gapi/include/opencv2/gapi/gopaque.hpp
@@ -325,7 +325,7 @@ namespace detail
 template<typename T> class GOpaque
 {
 public:
-    GAPI_OBJECT_META
+    GAPI_OBJECT(Meta)
     // Host type (or Flat type) - the type this GOpaque is actually
     // specified to.
     /// @private
