+++ modules/gapi/include/opencv2/gapi/gscalar.hpp
@@ -41,7 +41,7 @@ struct GOrigin;
 class GAPI_EXPORTS_W_SIMPLE GScalar
 {
 public:
-    GAPI_OBJECT_META
+    GAPI_OBJECT(Meta)
     /**
      * @brief Constructs an empty GScalar
      *
