+++ modules/gapi/include/opencv2/gapi/gstreaming.hpp
@@ -86,8 +86,9 @@ using GOptRunArgs = std::vector<GOptRunArg>;
 namespace detail {
 
 template<typename T> inline GOptRunArgP wrap_opt_arg(optional<T>& arg) {
-    static_assert(!cv::is_contain_tag<T, cv::tag::Meta>::value,
-                  "gin/gout must not be used with G* structures with tag::Meta");
+    static_assert(!(cv::gapi::has_tag<T, cv::gapi::tag::Meta>::value ||
+                    cv::gapi::has_tag<T, cv::gapi::tag::GraphRejected>::value),
+                  "gin/gout must not be used with G* structures with tag::Meta or gapi::own");
     // By default, T goes to an OpaqueRef. All other types are specialized
     return GOptRunArgP{OptionalOpaqueRef(arg)};
 }
